requirements.txt
Django>=5.2.6
djangorestframework>=3.15.0
drf-spectacular>=0.26.0
authlib>=1.3.0
pyjwt>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.0
